chore: rename node-test-helpers.js to node.js
