+++ integration/test-runner/tests/location-change/runLocationChangeTest.ts
@@ -4,7 +4,7 @@ import { runTests } from '@web/test-runner-core/test-helpers';
 import assert from 'node:assert/strict';
 import { before, describe, it } from 'node:test';
 import { resolve } from 'path';
-import { assertIncludes } from '../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes } from '../../../../test-helpers/node.js';
 
 export function runLocationChangeTest(
   config: Partial<TestRunnerCoreConfig> & { browsers: BrowserLauncher[] },
