+++ packages/dev-server-core/test/middleware/etagCacheMiddleware.test.ts
@@ -4,7 +4,7 @@ import assert from 'node:assert/strict';
 import { afterEach, beforeEach, describe, it } from 'node:test';
 import path from 'path';
 
-import { timeout } from '../../../../test-helpers/node-test-helpers.js';
+import { timeout } from '../../../../test-helpers/node.js';
 import type { DevServer } from '../../dist/server/DevServer.js';
 import { createTestServer } from '../helpers.ts';
