+++ packages/dev-server-core/test/middleware/serveFilesMiddleware.test.ts
@@ -2,7 +2,7 @@ import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 import path from 'path';
 
-import { assertIncludes } from '../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes } from '../../../../test-helpers/node.js';
 import { createTestServer } from '../helpers.ts';
 
 describe('serveFilesMiddleware', () => {
