+++ packages/dev-server-esbuild/test/target.test.ts
@@ -2,7 +2,7 @@ import { createTestServer } from '@web/dev-server-core/test-helpers';
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 
-import { assertIncludes } from '../../../test-helpers/node-test-helpers.js';
+import { assertIncludes } from '../../../test-helpers/node.js';
 import { esbuildPlugin } from '../dist/index.js';
 
 const modernJs = `
