+++ packages/dev-server-esbuild/test/tsx.test.ts
@@ -2,7 +2,7 @@ import { createTestServer } from '@web/dev-server-core/test-helpers';
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 
-import { assertIncludes, assertNotIncludes } from '../../../test-helpers/node-test-helpers.js';
+import { assertIncludes, assertNotIncludes } from '../../../test-helpers/node.js';
 import { esbuildPlugin } from '../dist/index.js';
 
 describe('esbuildPlugin TSX', { timeout: 5000 }, () => {
