+++ packages/dev-server-rollup/test/node/plugins/alias.test.ts
@@ -1,7 +1,7 @@
 import rollupAlias from '@rollup/plugin-alias';
 import { describe, it } from 'node:test';
 
-import { assertIncludes, fetchText } from '../../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes, fetchText } from '../../../../../test-helpers/node.js';
 import { fromRollup } from '../../../dist/fromRollup.js';
 import { createTestServer } from '../test-helpers.ts';
