+++ packages/dev-server-rollup/test/node/plugins/babel.test.ts
@@ -3,7 +3,7 @@ import rollupBabel from '@rollup/plugin-babel';
 import { createRequire } from 'node:module';
 import { describe, it } from 'node:test';
 
-import { assertIncludes, fetchText } from '../../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes, fetchText } from '../../../../../test-helpers/node.js';
 import { fromRollup } from '../../../dist/index.js';
 import { createTestServer } from '../test-helpers.ts';
