+++ packages/dev-server-rollup/test/node/plugins/postcss.test.ts
@@ -5,7 +5,7 @@ import { describe, it } from 'node:test';
 import { resolve } from 'path';
 import rollupPostcss from 'rollup-plugin-postcss';
 
-import { assertIncludes, fetchText } from '../../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes, fetchText } from '../../../../../test-helpers/node.js';
 import { fromRollup } from '../../../dist/index.js';
 import { createTestServer } from '../test-helpers.ts';
