+++ packages/dev-server-rollup/test/node/plugins/replace.test.ts
@@ -1,7 +1,7 @@
 import rollupReplace from '@rollup/plugin-replace';
 import { describe, it } from 'node:test';
 
-import { assertIncludes, fetchText } from '../../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes, fetchText } from '../../../../../test-helpers/node.js';
 import { fromRollup } from '../../../dist/index.js';
 import { createTestServer } from '../test-helpers.ts';
