+++ packages/rollup-plugin-html/test/src/output/css.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import { assertIncludes } from '../../../../../test-helpers/node-test-helpers.js';
+import { assertIncludes } from '../../../../../test-helpers/node.js';
 import {
   calculateRelativePath,
   createAssetPlaceholder,
